// sanity.types.ts
export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};

export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type BuildingsQueryResult = Array<{
  _id: string;
  status: "draft" | "published";
  title: string | "Sem t\xEDtulo";
  shortDescription: string | null;
  address: string | null;
  type: "comercial" | "residencial" | null;
  image: {
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    alt?: string;
    _type: "image";
  } | null;
  date: string;
}>;